Bug 1366097 - Part 5. Implement VectorImage::GetImageContainer and VectorImage::IsImageContainerAvailable.
This is largely trivial because the meat of the implementation is
located in ImageResource and we already added GetFrameInternal.
Interestingly VectorImage::IsUnlocked does not actually check if the
image is locked, but instead only checks for animation consumers. This
is consistent with its historical behavior on when to issue an unlocked
draw event.