Bug 1389021 - RemoteCompositorSession::Shutdown should call CompositorManagerChild::Shutdown. r=me
Historically CompositorBridgeChild::Destroy was sufficient to teardown
the IPC channel between a content process and the GPU process. This is
because PCompositorBridge was a top level protocol. However with the
introduction of PCompositorManager, destroying the CompositorBridgeChild
merely destroys that subprotocol, and does not touch the IPC channel
otherwise.

From the best results we have collected thus far, it appears the content
process is not properly releasing its CompositorManagerChild's IPC
channel on shutdown. This keeps the compositor thread alive in the GPU
process and triggers a shutdown hang. The content process should
automatically destroy the channel when it quick exits, but this may not
be the case (or at least, it may take longer to timeout than we expect).
As such, we are make the change here to down down the
CompositorManagerChild when the RemoteCompositorSession is torn down.

Additionally, changing the shutdown ordering of CompositorBridgeChild
revealed a bug where if CompositorBridgeChild::ActorDestroy was called
before Destroy, we would leak the CompositorBridgeChild due to a cycle
with the layout manager. Now in Destroy we ensure that cycle is broken
if ActorDestroy was called first to allow the cleanup of the object.