Bug 1398697 - Part 2. Make VectorImage use ImageResource::mHasSize and mSize to cache its size information.
While not an expensive operation, it is also not trivial to get the
width and height from an SVG. Since we will be potentially asking for
this information a lot more with enabling factor of 2 on SVGs, we
should start caching it and updating it if it becomes invalidated. This
will also become useful for image container support in the future.

The new behavior slightly differs from the original, in that if the size
changes, VectorImage::GetIntrinsicSize, GetHeight and GetWidth will
return the cached size until the invalidation is handled. This should be
okay because the surface remains cached as well, until the invalidation
is handled, thus remaining externally consistent.