Bug 1404422 - Part 1. Ensure we consistently add imgRequestProxy objects to the expected load group.
imgRequestProxy::PerformClone now adds the cloned request to the given
document's load group before requesting the notifications. We ensure
that any removals from the load group occur outside the current
execution context.

Legacy listeners may use imgRequestProxy::SyncClone to request
notifications on the image state. Ideally they would not, but they do
not work as expected with the asynchronous notifications all new callers
must use. While in theory this would suggest their code is re-entrant,
not all of it is. In particular we need to be sensitive about when we
remove a request from a load group.

Calling imgRequestProxy::SetLoadGroup is now officially unsupported. It
did not have a predictable effect and it appears to be unused. It is
somewhat complicated to support given we must be sensitive about what
context we execute removing the request from the original load group.