Bug 1408532 - Ensure ImageBridgeParent cleans up old actors if the process ID is reused.
On Windows, it sometimes reuses the process ID immediately after the
original process died. This has been observed with the GPU process in
particular, but could in theory happen with the content processes as
well.

Shutdown hangs have been observed where ImageBridgeParent objects are
left unfreed and/or unclosed. This should be impossible because
sImageBridges should contain a list of all existing ImageBridgeParent
objects which we force closed during shutdown. The only way that we can
forget about an object (that I could determine) is if the process ID is
reused and we didn't get an ImageBridgeParent::ActorDestroy call.

As such, this change checks for a previous ImageBridgeParent object in
the map with the same process ID before inserting the new one. It will
ensure the old object is properly destroyed and prevent any related
shutdown hangs.