Bug 1419912 - Part 1. Make SharedSurfacesChild delegate the decision to update its image keys.
SharedSurfacesChild previously based its decision to update its image
keys for each layer manager on the generation counter of the image
container the surface is stored in. With the advent of multiple image
containers per image, the same surface can be stored many times, each
with their own generation counter.

Ultimately whether or not the surface has changed, and thus needs a new
image key, is independent of the container it is stored in. The image
itself has a far better idea of when this actually happens, so this
patch refactors the decision making around the image key to delegate the
authority to force an image key update. A later patch in this series
will call SharedSurfacesChild::SurfaceUpdated at the appropriate time.