Bug 1446309 - Part 1. Add support to getting an image container only for fully decoded images.
imgIContainer::Draw returns a draw result from which callers have
reasoned about whether or not to display alternative content instead
(such as a previous image retained after a source change). With image
containers, there was no way to make a similar decision, and so we have
once again seen image flickering with WebRender.

This patch adds a new flag, FLAG_WANT_FULLY_DECODED, which is used by
GetImageContainer(AtSize). If given, it will only return image
containers which contain the fully decoded result. A caller can provide
this flag on the current image's call, and if it gets nothing, attempt
to fallback to the previous image.