Bug 1446309 - Part 2. Make nsDisplayImage support using the previous image with WebRender.
In order to avoid flickering, we retain the previous image. We will
preferentially draw with the previous image if the current image is not
fully decoded. This was not used with the WebRender path and thus
flickering has reappeared. This patch adds support to falling back to
the previous image.