Bug 1449559 - Add nsIURI::GetSpecHash to generate a hash from spec without a copy.
There are a few places (ImageCacheKey, nsURIHashKey) which generate a
hash from the URI spec, but only keep a temporary copy of the spec
retrieved via nsIURI::GetSpec. This patch adds a new method to nsIURI,
GetSpecHash, which generates a suitable hash from the spec, or from the
elements from which we compose the spec. It may not produce the same
result as using HashString directly on the result from GetSpec, but it
should be otherwise suitable for a similar purpose.