Bug 1453454 - SourceBuffer::Compact should avoid running on the main thread.
SourceBuffer::Compact is expensive if our initial allocation was not the
correct size for the received encoded image data. This can happen if the
HTTP response does not provide a Content-Length header (e.g. chunked
response), or it provided an incorrect Content-Length (bad web servers).
The majority of the time it will do nothing since the given value is
correct, but when it isn't, it will now defer the actual compacting to a
decoder thread over doing it immediately on the main thread.