Bug 1453747 - Stop rounding the image dest rect with WebRender.
We are incorrectly using the unrounded dest rect to calculate the image
decode size in ComputeImageContainerDrawingParameters, while passing the
rounded dest rect to WebRender. This mismatch could cause images to
sized slightly different than is optimal for WebRender. Using the
correct rect seems to allow us to remove the extra snapping logic added
to work around this mistake.