Bug 1465619 - Part 3. Improve the clear rect calculations to take into account recycling in BlendAnimationFilter.
The clear rect and the recycle rect can overlap, and depending on the
size of the clear rect, it could be a significant amount of data that
needs to be copied from the restore frame. This patch minimizes the
copying for a row which contains both the recycle rect and the clear
rect.