Bug 1452513 - Avoid issuing transactions with WebRender when the namespace has changed.
When the namespace changes (e.g. due to a tab move between windows), we
may get stale transaction requests that we need to ignore. In
WebRenderBridgeParent::RecvSetDisplayList, we would automatically send
any unsent transaction data when exiting the method, but this did not
take into account the staleness. This patch ensures we only flush the
data if we actually want it.

The transaction in question that was observed and causing crashes was
UpdateImageBuffer.