Bug 1446309 - Part 2. Make nsDisplayImage fallback to the previous image to avoid flickering.
When the current image for an nsImageFrame/nsDisplayImage is not yet
ready, we display the previous image, if any available, to avoid
flickering while we wait for decoding to finish. On the WebRender path,
this functionality was lost since we did not have the draw result
information with image containers. With the API updated in part 1, we
can now do this to avoid flickering.